chore: move janitor functions to janitor.py
