+++ sqlmesh/core/state_sync/__init__.py
@@ -20,5 +20,4 @@ from sqlmesh.core.state_sync.base import (
     Versions as Versions,
 )
 from sqlmesh.core.state_sync.cache import CachingStateSync as CachingStateSync
-from sqlmesh.core.state_sync.common import cleanup_expired_views as cleanup_expired_views
 from sqlmesh.core.state_sync.db import EngineAdapterStateSync as EngineAdapterStateSync
